Add functionality to resolve_schema for expressions2
