+++ src/python/schema.rs
@@ -3,18 +3,13 @@ use std::sync::Arc;
 use pyo3::prelude::*;
 
 use super::datatype::PyDataType;
+use super::field::PyField;
 use crate::datatypes;
-use crate::python::datatype;
 use crate::schema;
 
 #[pyclass]
 pub struct PySchema {
     pub schema: schema::SchemaRef,
-}
-
-#[pyclass]
-pub struct PyField {
-    pub field: datatypes::Field,
 }
 
 #[pymethods]
@@ -48,33 +43,6 @@ impl PySchema {
         Ok(PySchema {
             schema: schema.into(),
         })
-    }
-}
-
-#[pymethods]
-impl PyField {
-    pub fn name(&self) -> PyResult<String> {
-        Ok(self.field.name.clone())
-    }
-
-    pub fn dtype(&self) -> PyResult<datatype::PyDataType> {
-        Ok(self.field.dtype.clone().into())
-    }
-
-    pub fn eq(&self, other: &PyField) -> PyResult<bool> {
-        Ok(self.field.eq(&other.field))
-    }
-}
-
-impl From<datatypes::Field> for PyField {
-    fn from(field: datatypes::Field) -> Self {
-        PyField { field }
-    }
-}
-
-impl From<PyField> for datatypes::Field {
-    fn from(item: PyField) -> Self {
-        item.field
     }
 }
